[#658] Extract change password and delete account forms
Moved Change Password and Delete Account forms to individual components from the main Account view.

* Simplified `AccountIndex.vue` by removing inline form code.
* Improved code maintainability and readability.
* Each function is now encapsulated in its dedicated component.
